requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.4.0
python-dotenv==1.0.0
requests==2.27.1
functions-framework==3.4.0
rich==13.3.0
inquirer==2.10.0
pytest==8.2.2
requests-mock==1.10.0
